feat: Adds coin flip functionality
Adds the coin flip functionality to the webpage.

The flip action is triggered by pressing the "Flip Coin" button.

This change includes removing the inline `onclick` event
handler from the button in `index.html` and adding an event
listener in `script.js`. It also adds basic styling to the flip button.